build: permit PREFIX, MFEM_DIR overrides
This commit enables users to override values of PREFIX and MFEM_DIR at
the command line. The motivation for this change is to enable building
Laghos with different MFEM builds by defining variables at the command
line or in scripts. This feature avoids polluting the repository with
trivial changes to paths in the project makefile when testing
different MFEM configurations, and permits building MFEM out-of-tree.